Cleanup in unit GuiTasks
I realized it is not necessary to use DisableTaskWindows/EnableTaskWindows in TGuiThread.Wait(). When an action shows a dialog from inside the message loop (via the hook), the situation is the same as with a standard windows timer: Its up to the code in the action to prevent issues due to nested message loops, exactly as for timer events.